Create mdb-mylock.py
This script will continuously run the query to fetch lock information from the performance_schema.data_locks table and display the results. The LOCK_TYPE filter is set to Record so that the script only displays the record locks. The script will run every 2 seconds to get the updated lock information.